Fix trailers race condition
See branch reproduce/unknown-status-issue.

We assume that trailers only responses are not an issue, and we should
never hang when there is no body in the response. This is consistent
with expectations of HTTP2, so any hangs would probably indicate an
issue with the underlying HTTP2 library.

Should be tested with the etcd example.