enforce consistency of request and response modes
- This adds type safety to ensure the RPCs are called with the
  expected protocol (e.g. you cannot call a unary rpc with a
  server_streaming entry point, etc.).

On the ocaml-protoc-plugin side, currently there are no markers for
the rpc modes - this interface will permit adding them in the future
without user facing changes.

On the ocaml-protoc plugin, the value mode flows from the proto file
definition and is checked in the user code as expected.

Implementation note: There's perhaps a way to shorten the mapping of
value-modes but I couldn't find one given that `Grpc` cannot depend on
`Ocaml_protoc`, and thus the `Value_mode` types are not equal.