fix the service names used by Typed_rpc
In the existing examples, the service name is separated from the
package name by a dot, which I inadvertently omitted in the previous
implementation.

Note that as long as the service name used by a client and a server is
the same, the right handler is executed, so there's some leeway in the
actual choice of the convention to use. The hope is that the dot
separated one is standard.