Sort dependencies in dune-project
This makes it is easier to avoid duplicated entries as well as making
it obvious were new dependencies need to be added, which makes it more
straight forward to write PRs that add new dependencies.

In the process I discovered that the notty dependency was duplicated
in `grpc-bench.opam`, which I fixed. This had no incidence on the
build.

The canonical order I used here was to put `ocaml` first, and then the
rest of dependencies alphabetically.